Move Component to Separate File
